refactor: hooksディレクトリをcomponentsディレクトリの外に配置するように修正
