+++ src/hooks/directories/parseTextToDirectories.test.ts
