+++ src/hooks/directories/resolveFullDirectoryPath.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import type { Directory } from "../../../types/directory";
+import type { Directory } from "../../types/directory";
 import { resolveFullDirectoryPath } from "./resolveFullDirectoryPath";
 
 describe("resolveFullDirectoryPath", () => {
