+++ src/hooks/directories/resolveFullDirectoryPath.ts
@@ -1,4 +1,4 @@
-import type { Directory } from "../../../types/directory";
+import type { Directory } from "../../types/directory";
 
 export const resolveFullDirectoryPath = (
 	directories: Directory[],
