+++ src/hooks/formatTextAsDirectoryTree/calculateDepth.test.ts
