+++ src/hooks/formatTextAsDirectoryTree/formatTextAsDirectoryTree.ts
