+++ src/hooks/formatTextAsDirectoryTree/generatePrefix.test.ts
