+++ src/hooks/formatTextAsDirectoryTree/getLines.ts
