+++ src/hooks/formatTextAsDirectoryTree/isLast.test.ts
