+++ src/hooks/formatTextAsDirectoryTree/separateLinses.test.ts
